Create Trouble
At first I wasn't getting the correct outputs. originally I had my variable  'discount' be added onto inside the if statements. 
discount += .30
and so on. It wasn't working out until I initialized it into one and instead had it subtract the discounts like that and continue multiplying. which is when i had to add the extra 1-discount at line 56 to fix that output. Not sure why I had to do that but it works.