Making Multiple Comparisons in Decision Statements
This one wasn't too bad. For a second or two I had to read the code again to see if the code already gave me a way to output the eligibility// not sure if that's a word// but I ended up adding an extra string variable to carry over my final decision into the output pretty easily I think.