Update examples
tldr;
num%10  // reads only the last digit
num/10   // actually takes it off the whole numberCode of adding up the sum of user input. 

Finding the remainder of sum/10 with an integer, it takes the last number and puts it wherever you want it to go.
With 953 % 10 3 would be the remainder. 95 would be the answer of 953/10 when it is an int, because it can not hold decimal values. 
After finding the last digit of the number, we need to actually take the last digit off the number to keep the loop moving and finding the rest of the digit sums.