Example of counter-controlled loops.
In a while loop here you see at the end of the loop you are adding to the variable //numberCounter// by an increment of 1 to keep it moving and looping until it hits the limit set in the while statement in line 30. More can be found in the seperate thread of the full copy of this code called "COUNTER-CONTROLLED".