How to do a sentinel control loop
Here we learn how to use a while loop when it's looking for something it shouldn't be. Which will end the loop. 
We forgot to put an input INSIDE of the loop, causing it be an infinite loop and ultimately not outputting what we wanted, the average amount of stars input ed. 
The totalStars value kept the total value of all the stars put in to be divided and have the average found. 
numPatrons variable kept the number of times we had input within the loop, that way we know what number to divide the totalStars value to find the average. It works. 

One thing though is that it was supposed to be any negative number but I wasn't sure how to do that so I was lazy and just put -1 knowing that would work in the end.