Hard level
Mostly because of the wording this problem was pretty difficult. I couldn't figure out for the longest the problem was saying //between// two input integers and not just the two input variables being multiplied and THEN finding the multiples. It was literally inbetween the two variables. Like the 9 numbers in between 1 and 10. 
```````````````````````````````````````````````````````````````
Write a program that prompts the user to enter two integers.

The program outputs how many numbers are multiples of 3 and how many numbers are multiples of 5 between the two integers (inclusive).