Nested Loops E
Kind of difficult. Really trains the before hand stuff of working out a problem. Here we have to look for simple patters within the letter to know where to put the stars. We can't just cout it out, that would be too easy and not effective in bigger projects. 

Most important thing I learned from this is the internet helps a lot, but there's a TON of garbage to go through. and how to use nested loops of course. 

May seem abstract but the use of nesting stuff seems to be a real building block for short patterns that repeated for a desired outcome. Like making the letter E out of stars.