strings
The only thing that gave me trouble were the strings, I'm not sure if I can jus put any letter as a variable in the string leftOrRight but if I didn't declared the L, R, X then the program would not run. Otherwise for what it was supposed to teach me, accumulating totals in this manner was not difficult and I think I have a greater grasp on loops now.