Forgot to add the prompt
In this lab, you use what you have learned about parallel arrays to complete a partially completed C++ program. The program should:

Either print the name and price for a coffee add-in from the Jumpin’ Jive Coffee Shop
Or it should print the message Sorry, we do not carry that.
Read the problem description carefully before you begin. The file provided for this lab includes the necessary variable declarations and input statements. You need to write the part of the program that searches for the name of the coffee add-in(s) and either prints the name and price of the add-in or prints the error message if the add-in is not found. Comments in the code tell you where to write your statements.

Instructions
Study the prewritten code to make sure you understand it.
Write the code that searches the array for the name of the add-in ordered by the customer.
Write the code that prints the name and price of the add-in or the error message, and then write the code that prints the cost of the total order.
Execute the program by clicking the Run button at the bottom of the screen. Use the following data:

Cream

Caramel

Whiskey

chocolate

Chocolate

Cinnamon

Vanilla