NOT MY CODE
Got lazy, and honestly do not understand how it figures out it is a prime or not. Is it checking the possible numbers or something? Idk. Here is the explanation though. 
---------------------------------------------------------------------------------------
This program takes a positive integer from user and stores it in variable n.

Then, for loop is executed which checks whether the number entered by user is perfectly divisible by i or not.
The for loop initiates with an initial value of i equals to 2 and increasing the value of i in each iteration.

If the number entered by user is perfectly divisible by i then, isPrime is set to false and the number will not be a prime number.

But, if the number is not perfectly divisible by i until test condition i <= n/2 is true means, it is only divisible by 1 and that number itself.

So, the given number is a prime number.
---------------------------------------------------------------------------------------
I think I understood it better now that I reread it, it cuts the number in half and checks. Still though a big ass number would probably make that take a while right? Or is it insignificant for a computer doing that type of math?