How to use a value returning function
Using the data type function bool. I had to put the upper and lower case version of the letters i was looking for making it a little longer but it did the job. Putting the expression I'm  evaluating within the return statement also made it a lot cleaner than using an if statement that I was originally using.  
Remember to declare the function before the main( ) of the program 
I was also trying to assign the bool value to an int y to display the 1 or 0 when I could just put it in the output as the function call. //Line 13//